(PC-29546)[API] refactor: move synchronize_accessibility_with_acceslibre command to script
